requirements.txt
Django
djangorestframework
django-rest-swagger
django-filter
django-cors-headers
django-model-utils
django-versatileimagefield
django-picklefield
djoser
markdown
requests
Pillow
python-dotenv
python-magic
drf-writable-nested
django-constance[databese]
psycopg2-binary
django-ckeditor
